REF: Remove `interop.internal` module
