--- openff/interchange/_tests/interoperability_tests/internal/__init__.py
@@ -1,1 +1,0 @@
-"""Interoperability tests for internal writers."""
