+++ openff/interchange/_tests/interoperability_tests/lammps/export/__init__.py
@@ -1,0 +1,1 @@
+"""Tests of LAMMPS exporter."""
