--- openff/interchange/_tests/unit_tests/interop/internal/__init__.py
@@ -1,1 +1,0 @@
-"""Unit tests for internal writers."""
