--- openff/interchange/interop/internal/__init__.py
@@ -1,3 +1,0 @@
-"""
-Custom functions for exporting from Interchange objects to molecular simulation files.
-"""
