--- openff/interchange/interop/internal/amber.py
@@ -1,3 +1,0 @@
-"""Old import paths for Amber interoperability. See openff.interchange.interop.amber."""
-
-from openff.interchange.interop.amber.export._export import to_inpcrd, to_prmtop
