+++ openff/interchange/interop/lammps/export/__init__.py
@@ -1,0 +1,3 @@
+"""Export to LAMMPS."""
+
+from openff.interchange.interop.lammps.export.export import to_lammps
