+++ openff/interchange/models.py
@@ -249,7 +249,7 @@ class PotentialKey(DefaultModel):
     .. code-block:: pycon
 
         >>> from openff.interchange.models import PotentialKey
-        >>> from openff.toolkit.typing.engines.smirnoff import ForceField
+        >>> from openff.toolkit import ForceField
         >>> parsley = ForceField("openff-1.0.0.offxml")
         >>> param = parsley["Bonds"].get_parameter({"id": "b55"})[0]
         >>> bond_55 = PotentialKey(id=param.smirks)
