+++ setup.cfg
@@ -15,7 +15,7 @@ max-line-length = 119
 ignore = E203,B028
 per-file-ignores =
     openff/interchange/_tests/unit_tests/test_types.py:F821
-    openff/interchange/smirnoff/__init__.py:F401
+    openff/interchange/**/__init__.py:F401
     openff/interchange/smirnoff/_nonbonded.py:F821
     openff/interchange/smirnoff/_gbsa.py:F821
     openff/interchange/smirnoff/_virtual_sites.py:F821
@@ -24,10 +24,6 @@ per-file-ignores =
     openff/interchange/components/*.py:F821
     openff/interchange/components/foyer.py:F401
     openff/interchange/components/_packmol.py:W503
-    openff/interchange/__init__.py:F401
-    openff/interchange/interop/internal/amber.py:F401
-    openff/interchange/interop/gromacs/__init__.py:F401
-    openff/interchange/interop/openmm/_import/__init__.py:F401
     openff/interchange/_tests/data/*:INP001
     plugins/*:INP001
